please read
please check the dates on the commits from branch "task-7-combat", they were done at the time of submission but the branch was not merged until now because the task was technically incomplete. I didn't think to merge because I thought the commits would be visible from the github page, but they aren't immediately visible.